feat: Introduce comprehensive transaction management with a new `Manage` page, `TransactionsContext`, `ThemeContext`, and dedicated tabs for overview, transactions, analytics, and settings.
