+++ src/App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import { AuthProvider } from '@/contexts/AuthContext';
 import { ThemeProvider } from '@/contexts/ThemeContext';
+import { TransactionsProvider } from '@/contexts/TransactionsContext';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { Toaster } from '@/components/ui/toaster';
 import { ProtectedRoute } from '@/components/ProtectedRoute';
@@ -38,90 +39,92 @@ function App() {
     <QueryClientProvider client={queryClient}>
       <AuthProvider>
         <ThemeProvider>
-          <Router>
-            <Routes>
-              <Route path="/" element={<AuthPage />} />
-              <Route path="/forgot-password" element={<ForgotPassword />} />
-              <Route path="/reset-password" element={<ResetPassword />} />
+          <TransactionsProvider>
+            <Router>
+              <Routes>
+                <Route path="/" element={<AuthPage />} />
+                <Route path="/forgot-password" element={<ForgotPassword />} />
+                <Route path="/reset-password" element={<ResetPassword />} />
 
-              {/* Protected User Routes */}
-              <Route
-                path="/dashboard"
-                element={
-                  <ProtectedRoute>
-                    <Dashboard />
-                  </ProtectedRoute>
-                }
-              />
-              <Route
-                path="/manage"
-                element={
-                  <ProtectedRoute>
-                    <Manage />
-                  </ProtectedRoute>
-                }
-              />
-              <Route
-                path="/insights"
-                element={
-                  <ProtectedRoute>
-                    <Insights />
-                  </ProtectedRoute>
-                }
-              />
-              <Route
-                path="/insights/:slug"
-                element={
-                  <ProtectedRoute>
-                    <InsightsPost />
-                  </ProtectedRoute>
-                }
-              />
+                {/* Protected User Routes */}
+                <Route
+                  path="/dashboard"
+                  element={
+                    <ProtectedRoute>
+                      <Dashboard />
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/manage"
+                  element={
+                    <ProtectedRoute>
+                      <Manage />
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/insights"
+                  element={
+                    <ProtectedRoute>
+                      <Insights />
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/insights/:slug"
+                  element={
+                    <ProtectedRoute>
+                      <InsightsPost />
+                    </ProtectedRoute>
+                  }
+                />
 
-              {/* Super Admin Routes */}
-              <Route
-                path="/superadmin"
-                element={
-                  <ProtectedRoute>
-                    <SuperAdminLayout>
-                      <SuperAdminDashboard />
-                    </SuperAdminLayout>
-                  </ProtectedRoute>
-                }
-              />
-              <Route
-                path="/superadmin/users"
-                element={
-                  <ProtectedRoute>
-                    <SuperAdminLayout>
-                      <UsersPage />
-                    </SuperAdminLayout>
-                  </ProtectedRoute>
-                }
-              />
-              <Route
-                path="/superadmin/content"
-                element={
-                  <ProtectedRoute>
-                    <SuperAdminLayout>
-                      <ContentPage />
-                    </SuperAdminLayout>
-                  </ProtectedRoute>
-                }
-              />
-              <Route
-                path="/superadmin/settings"
-                element={
-                  <ProtectedRoute>
-                    <SuperAdminLayout>
-                      <SettingsPage />
-                    </SuperAdminLayout>
-                  </ProtectedRoute>
-                }
-              />
-            </Routes>
-            <Toaster />
-          </Router>
+                {/* Super Admin Routes */}
+                <Route
+                  path="/superadmin"
+                  element={
+                    <ProtectedRoute>
+                      <SuperAdminLayout>
+                        <SuperAdminDashboard />
+                      </SuperAdminLayout>
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/superadmin/users"
+                  element={
+                    <ProtectedRoute>
+                      <SuperAdminLayout>
+                        <UsersPage />
+                      </SuperAdminLayout>
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/superadmin/content"
+                  element={
+                    <ProtectedRoute>
+                      <SuperAdminLayout>
+                        <ContentPage />
+                      </SuperAdminLayout>
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/superadmin/settings"
+                  element={
+                    <ProtectedRoute>
+                      <SuperAdminLayout>
+                        <SettingsPage />
+                      </SuperAdminLayout>
+                    </ProtectedRoute>
+                  }
+                />
+              </Routes>
+              <Toaster />
+            </Router>
+          </TransactionsProvider>
         </ThemeProvider>
       </AuthProvider>
     </QueryClientProvider>
